Hallway generation now works entirely\!

+++ src/PathfindingTile.java
@@ -7,11 +7,11 @@ public class PathfindingTile {
 	public PathfindingTile parent; // Tile's parent, used for finding final path.
 	public int x;
 	public int y;
-	public int f; // g + h
-	public int g; // Guess, 10 for straight, 14 for diagnol.
-	public int h; // Heuristic (Manahattan), g * abs(curX - targetX) + abs(curY - targetY)
+	public double f; // g + h
+	public double g; // Guess, 10 for straight, 14 for diagnol.
+	public double h; // Heuristic (Manahattan), g * abs(curX - targetX) + abs(curY - targetY)
 
-	public PathfindingTile(PathfindingTile parent, int x, int y, int f, int g, int h) {
+	public PathfindingTile(PathfindingTile parent, int x, int y, double f, double g, double h) {
 		this.parent = parent;
 		this.x = x;
 		this.y = y;
